# openapi.json
{
  "components": {
    "schemas": {
...
      "ErrorDetail": {
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "fields": {
            "items": {
              "$ref": "#/components/schemas/ErrorField"
            },
            "title": "Fields",
            "type": "array"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "request_id": {
            "title": "Request Id",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message",
          "request_id"
        ],
        "title": "ErrorDetail",
        "type": "object"
      },
      "ErrorField": {
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "field": {
            "title": "Field",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "field",
          "code",
          "message"
        ],
        "title": "ErrorField",
        "type": "object"
      },
      "ErrorResponse": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
      "ExpertCommentResponse": {
        "properties": {
          "author_label": {
            "title": "Author Label",
            "type": "string"
          },
          "content": {
            "title": "Content",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          }
        },
        "required": [
          "id",
          "author_label",
          "content"
        ],
        "title": "ExpertCommentResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ProductCategoryListResponse": {
        "examples": [
          {
            "items": [
              {
                "name": "전체",
                "slug": "all"
              },
              {
                "name": "종합비타민",
                "slug": "multivitamin"
              },
              {
                "name": "비타민B군",
                "slug": "vitamin-b"
              },
              {
                "name": "비타민C",
                "slug": "vitamin-c"
              },
              {
                "name": "오메가3",
                "slug": "omega-3"
              }
            ]
          }
        ],
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ProductCategoryResponseItem"
            },
            "title": "Items",
            "type": "array"
          }
        },
        "required": [
          "items"
        ],
        "title": "ProductCategoryListResponse",
        "type": "object"
      },
      "ProductCategoryResponseItem": {
        "properties": {
          "name": {
            "title": "Name",
            "type": "string"
          },
          "slug": {
            "title": "Slug",
            "type": "string"
          }
        },
        "required": [
          "slug",
          "name"
        ],
        "title": "ProductCategoryResponseItem",
        "type": "object"
      },
      "ProductDetailResponse": {
        "examples": [
          {
            "brand": "고려은단",
            "category_slugs": [
              "multivitamin"
            ],
            "currency": "KRW",
            "display_price": 0,
            "expert_comments": [
              {
                "author_label": "MJ's COMMENT",
                "content": "종합비타민 카테고리 안내와 섭취 주의사항",
                "id": "24000000-0000-4000-8000-000000000101"
              }
            ],
            "id": "22000000-0000-4000-8000-000000000101",
            "image_url": "/static/products/catalog-placeholder.svg",
            "name": "고려은단 멀티비타민 올인원",
            "nutrients": [
              {
                "amount_per_unit": "100",
                "code": "VITAMIN_C",
                "name": "비타민 C",
                "unit": "MG"
              }
            ],
            "package": {
              "unit_form": "TABLET",
              "units_per_package": "60"
            },
            "product_type": "SUPPLEMENT",
            "sku": "KORYO-EUNDAN-MULTIVITAMIN-ALL-IN-ONE"
          }
        ],
        "properties": {
          "brand": {
            "title": "Brand",
            "type": "string"
          },
          "category_slugs": {
            "items": {
              "type": "string"
            },
            "title": "Category Slugs",
            "type": "array"
          },
          "currency": {
            "const": "KRW",
            "default": "KRW",
            "title": "Currency",
            "type": "string"
          },
          "display_price": {
            "title": "Display Price",
            "type": "integer"
          },
          "expert_comments": {
            "items": {
              "$ref": "#/components/schemas/ExpertCommentResponse"
            },
            "title": "Expert Comments",
            "type": "array"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "image_url": {
            "title": "Image Url",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "nutrients": {
            "items": {
              "$ref": "#/components/schemas/ProductNutrientResponse"
            },
            "title": "Nutrients",
            "type": "array"
          },
          "package": {
            "$ref": "#/components/schemas/ProductPackageResponse"
          },
          "product_type": {
            "title": "Product Type",
            "type": "string"
          },
          "sku": {
            "title": "Sku",
            "type": "string"
          }
        },
        "required": [
          "id",
          "sku",
          "product_type",
          "brand",
          "name",
          "image_url",
          "display_price",
          "category_slugs",
          "package",
          "nutrients",
          "expert_comments"
        ],
        "title": "ProductDetailResponse",
        "type": "object"
      },
...
      "ProductNutrientResponse": {
        "properties": {
          "amount_per_unit": {
            "pattern": "^(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "title": "Amount Per Unit",
            "type": "string"
          },
          "code": {
            "title": "Code",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "unit": {
            "enum": [
              "MG",
              "G",
              "MCG",
              "IU"
            ],
            "title": "Unit",
            "type": "string"
          }
        },
        "required": [
          "code",
          "name",
          "amount_per_unit",
          "unit"
        ],
        "title": "ProductNutrientResponse",
        "type": "object"
      },
      "ProductPackageResponse": {
        "properties": {
          "unit_form": {
            "enum": [
              "TABLET",
              "CAPSULE",
              "SCOOP",
              "PACKET"
            ],
            "title": "Unit Form",
            "type": "string"
          },
          "units_per_package": {
            "pattern": "^(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "title": "Units Per Package",
            "type": "string"
          }
        },
        "required": [
          "unit_form",
          "units_per_package"
        ],
        "title": "ProductPackageResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "AccessCookieAuth": {
        "description": "HttpOnly access JWT 쿠키",
        "in": "cookie",
        "name": "yeongyangkkuk_access_token",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "영양꾹 1차 로컬 MVP 백엔드 API",
    "title": "영양꾹 백엔드",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/curation/categories": {
      "get": {
        "operationId": "curation_list_product_categories",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductCategoryListResponse"
                }
              }
            },
            "description": "가상 전체와 활성 제품 카테고리 목록"
          },
          "503": {
            "content": {
              "application/json": {
                "example": {
                  "error": {
                    "code": "SERVICE_UNAVAILABLE",
                    "fields": [],
                    "message": "서비스가 아직 준비되지 않았습니다.",
                    "request_id": "opaque-request-id"
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "PostgreSQL 조회 실패"
          }
        },
        "summary": "활성 제품 카테고리 조회",
        "tags": [
          "큐레이션"
        ]
      }
    },
...
    "/api/v1/curation/products/{product_id}": {
      "get": {
        "operationId": "curation_get_recommended_product",
        "parameters": [
          {
            "in": "path",
            "name": "product_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Product Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductDetailResponse"
                }
              }
            },
            "description": "게시 추천 제품 상세"
          },
          "404": {
            "content": {
              "application/json": {
                "example": {
                  "error": {
                    "code": "PRODUCT_NOT_FOUND",
                    "fields": [],
                    "message": "제품을 찾을 수 없습니다.",
                    "request_id": "opaque-request-id"
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "공개 추천 제품 없음"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "503": {
            "content": {
              "application/json": {
                "example": {
                  "error": {
                    "code": "SERVICE_UNAVAILABLE",
                    "fields": [],
                    "message": "서비스가 아직 준비되지 않았습니다.",
                    "request_id": "opaque-request-id"
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "PostgreSQL 조회 실패"
          }
        },
        "summary": "게시 추천 제품 상세 조회",
        "tags": [
          "큐레이션"
        ]
      }
    },
    "/api/v1/curation/products/{product_id}/purchase": {
      "get": {
        "operationId": "curation_redirect_to_purchase",
        "parameters": [
          {
            "in": "path",
            "name": "product_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Product Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "307": {
            "description": "활성 외부 구매 링크로 임시 이동",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string"
                }
              },
              "Location": {
                "schema": {
                  "format": "uri",
                  "type": "string"
                }
              },
              "Referrer-Policy": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "examples": {
                  "product_not_found": {
                    "summary": "공개 추천 제품 없음",
                    "value": {
                      "error": {
                        "code": "PRODUCT_NOT_FOUND",
                        "fields": [],
                        "message": "제품을 찾을 수 없습니다.",
                        "request_id": "opaque-request-id"
                      }
                    }
                  },
                  "purchase_link_not_found": {
                    "summary": "활성 구매 링크 없음",
                    "value": {
                      "error": {
                        "code": "PURCHASE_LINK_NOT_FOUND",
                        "fields": [],
                        "message": "구매 링크를 찾을 수 없습니다.",
                        "request_id": "opaque-request-id"
                      }
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "공개 추천 제품 또는 활성 구매 링크 없음"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "503": {
            "content": {
              "application/json": {
                "example": {
                  "error": {
                    "code": "SERVICE_UNAVAILABLE",
                    "fields": [],
                    "message": "서비스가 아직 준비되지 않았습니다.",
                    "request_id": "opaque-request-id"
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "PostgreSQL 조회 또는 구매 URL 검증 실패"
          }
        },
        "summary": "활성 외부 구매 링크로 이동",
        "tags": [
          "큐레이션"
        ]
      }
    },
...
  }
}